Make Error work work with and without cgo enabled
Let's say you have the following code:

	// ErrUnique reports if this error reports a UNIQUE constraint violation.V
	func ErrUnique(err error) bool {
		var sqlErr *sqlite3.Error
		if errors.As(err, &sqlErr) && sqlErr.ExtendedCode == sqlite3.ErrConstraintUnique {
			return true
		}
		var pqErr *pq.Error
		if errors.As(err, &pqErr) && pqErr.Code == "23505" {
			return true
		}
		return false
	}

This doesn't work with CGO_ENABLED=0 since:

	$ CGO_ENABLED=0 go test .
	# zgo.at/zdb [zgo.at/zdb.test]
	./zdb.go:16:14: undefined: sqlite3.Error
	./zdb.go:17:55: undefined: sqlite3.ErrConstraintUnique

Now, it's not too hard to create two implementations of this: a cgo version
which includes the SQLite and PostgreSQL check, and a non-cgo version which
includes just the PostgreSQL check, but it's all a bit cumbersome and requires
duplicate code or wrapper functions and more indirection than really needed.

Since there are just two methods that actually use the C "package", I think it
makes sense to put those in a error_cgo.go file, and put some basic shims in the
static_mock.go file – this way the code always compiles.
