+++ error.go
@@ -1,25 +1,16 @@
-// Copyright (C) 2019 Yasuhiro Matsumoto <mattn.jp@gmail.com>.
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file.
-
 package sqlite3
 
-/*
-#ifndef USE_LIBSQLITE3
-#include <sqlite3-binding.h>
-#else
-#include <sqlite3.h>
-#endif
-*/
-import "C"
 import "syscall"
+
+// Make sure the error interface is implemented.
+var (
+	_ error = Error{}
+	_ error = ErrNo(0)
+	_ error = ErrNoExtended(0)
+)
 
 // ErrNo inherit errno.
 type ErrNo int
-
-// ErrNoMask is mask code.
-const ErrNoMask C.int = 0xff
 
 // ErrNoExtended is extended errno.
 type ErrNoExtended int
@@ -73,24 +64,6 @@ func (err ErrNo) Error() string {
 // Extend return extended errno.
 func (err ErrNo) Extend(by int) ErrNoExtended {
 	return ErrNoExtended(int(err) | (by << 8))
-}
-
-// Error return error message that is extended code.
-func (err ErrNoExtended) Error() string {
-	return Error{Code: ErrNo(C.int(err) & ErrNoMask), ExtendedCode: err}.Error()
-}
-
-func (err Error) Error() string {
-	var str string
-	if err.err != "" {
-		str = err.err
-	} else {
-		str = C.GoString(C.sqlite3_errstr(C.int(err.Code)))
-	}
-	if err.SystemErrno != 0 {
-		str += ": " + err.SystemErrno.Error()
-	}
-	return str
 }
 
 // result codes from http://www.sqlite.org/c3ref/c_abort_rollback.html
