+++ static_mock.go
@@ -35,3 +35,6 @@ func (c *SQLiteConn) RegisterCommitHook(func() int)                            {
 func (c *SQLiteConn) RegisterFunc(string, interface{}, bool) error             { return errorMsg }
 func (c *SQLiteConn) RegisterRollbackHook(func())                              {}
 func (c *SQLiteConn) RegisterUpdateHook(func(int, string, string, int64))      {}
+
+func (err Error) Error() string         { return "no error message in CGO_ENABLED=0" }
+func (err ErrNoExtended) Error() string { return Error{}.Error() }
